requirements.txt
polars
pycryptodome
pyyaml
